+++ Assets/Prefabs/Guns/Machinegun.prefab
@@ -850,7 +850,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   gunConfiguration: {fileID: 11400000, guid: addb4062aacc4ee45aeb2881f5fc2f05, type: 2}
-  magazine: {fileID: 1110103690579164734}
+  animationObject: {fileID: 1110103690579164734}
   reloadAnimationClip: {fileID: 7400000, guid: efc332ec4a6cfff4b9c4a7ae33b730d9, type: 2}
   impactType: {fileID: 11400000, guid: 4a8931dfd9cfcb94dae2060677c798da, type: 2}
 --- !u!1001 &6507464826908258554
