+++ Assets/Prefabs/Guns/Pistol.prefab
@@ -902,7 +902,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   gunConfiguration: {fileID: 11400000, guid: 6ab3fb35dce4c4645848a01a9d941a75, type: 2}
-  magazine: {fileID: 2146904259634683510}
+  animationObject: {fileID: 2146904259634683510}
   reloadAnimationClip: {fileID: 7400000, guid: 503f3e49595caf7448aed2176c19aef6, type: 2}
   impactType: {fileID: 11400000, guid: 4a8931dfd9cfcb94dae2060677c798da, type: 2}
 --- !u!1001 &2561238631588639659
