+++ Assets/Prefabs/Player Backup/Player Backup - 1.prefab
@@ -4485,7 +4485,7 @@ GameObject:
   - component: {fileID: 6546904019569828687}
   - component: {fileID: 6018052358273083305}
   m_Layer: 8
-  m_Name: Player Backup
+  m_Name: Player Backup - 1
   m_TagString: Player
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
