+++ Assets/Prefabs/Player Backup/Player.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 4744cc50bfb38714a9ea7e4be0fbe0c9
-folderAsset: yes
-DefaultImporter:
+guid: f77dbbe1fad261b46948e123c670673c
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
