+++ Assets/Scripts/Weapon System/Animation/GunHitscanAnimations.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 421b4e3fd0f4e994ea884815fb65e549
+guid: fc6152a479b88aa4eaddcbbf24155ec9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
